backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cached-property==1.5.2
certifi==2025.6.15
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.103.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
openai==1.39.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
